Use $_SERVER['DOCUMENT_ROOT'] instead of $DOCUMENT_ROOT in the slow lookup function
Also use this for loading the sqlite DB, instead of a
relative path value

Reformat the sqlite lookup function to be indented
with spaces instead of tabs

Add many comments to this part

Add prefix detection code, so /function.echo type of
lookups will work with the sqlite lookup method too. This
hopefully fixes #23652.

Remove ending ?> for consistency

Changes tested on my local mirror with sqlite and without it