Use divs to present the TOC items on manual pages on the left side of the window
Add images according to php.net rules to these
two styles (toca, toci) depending on what mirror
they are used on

Also rewrite the TOC generator code a bit, so it
identifies items properly when their title and titleabbrev
differ [was a bug on many pages]

Also wrap too long function names, so we have the
same amount of content and nav area on pages with
long function names in the TOC [eg. xml functions]

The new div based approach means that we have *much*
less HTML to output, and the layout of the menu is better
than before, as the box images are aligned to the text.
Tested with Mozilla and Konqueror. Apart from pages with
long function names in the TOC, this should look very similar
to the display of the TOC before