Implemented search auto-complete.
The auto-completing search feature uses the popular
jquery.autocomplete plugin:

 http://docs.jquery.com/Plugins/Autocomplete

The auto-complete data set is held client-side. This
incurs one http request to load the search index when
the page is loaded (via jQuery.getScript()), thereby 
avoiding additional requests as the user types in the
search field. This makes the auto-complete feature 
very responsive.

The search index file defines a 'searchIndex' array 
containing the name and page (filename) of each entry 
in the manual. The file is 382 KB uncompressed. Gzip
compression (via the cli) reduces this to 53 KB. Proper
configuration of server-side compression and expire 
headers should make this download relatively painless.

The search index file resides under the js folder 
and is named: 'search-index-<lang>.js'. There is only
one index file at this time ('en') and it was manually
generated.

The auto-complete feature has one major short-coming;
the results are not sorted/ranked by relevance. This
will required modification of the auto-complete 
plugin or a server-side suggest script to overcome.