Merged styles from structure.css into theme.css and deleted structure.css.
I suspect that the intent of having these two files was to separate positional
styles from colors and fonts, etc. In practice, this separation was blurry at 
best and made it much more difficult to manage styles for a given dom element 
because they were spread across two files.

Hopefully the fallout from this change is minimal.