Reset HTML5 elements in the reset stylesheet by adding them to the master element list and adding a display: block rule for the block level elements.
The original plan here was to replace Eric Meyer's reset stylesheet with
Richard Clark's HTML5 reset stylesheet, but some of the changes he's made
aren't really ones we want, such as the different handling of list-style by
default, not resetting text-decoration on anchors, and removing a couple of
deprecated elements that we're still using (hello <acronym>). Since we'd have
to modify whichever reset stylesheet we use, I've stuck with the old one as the
base for now, since we already know it works with the site.