Tinker with the positioning of the what-is-php block on the home page.
With two releases, this should be identical to the previous positioning. With
more than two releases, the text is vertically aligned, then moved ever so
slightly up to avoid the feeling of the chunky Begin Tutorial button being too
far off centre.

This has mostly been done by feel, so everyone else may find this horrible, but
I think it's an improvement on the previous iteration where that block was just
at the top.