Restore the old site's default search, which is to try the quickref first.
This is actually an improvement on the old site, really, because the previous
case of "all php.net sites" is handled by the last item in the dropdown. My
inclination for now is to not worry about the other search modes that were
supported in the old site's search bar: many of them are for subsites that have
their own search, such as bugs.php.net and pecl.php.net, and the others have
tabs on results.php anyway.