Download system tweaks
Remove reduntant/duplicate code between do-download and get-download

Also make this page more interesting when the download is missing from
this specific mirror by give the choise of going to other mirrors to check.
Rather then blindly go to the download file though we go to the same page
on the other mirror so we can check if the file exists there before giving
the link to the actual real file.

This also makes it a lot easier to work with on developers checkouts