Optimizes Usergroup-display
This comit adds three things:

1. Support for 2-letter Countrycodes as defined in ISO 3166-1 alpha2.
For that I added a list of country codes from
http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Officially_assigned_code_elements

2. Caching of retrieved usergroup-information. For that the list of
usergroups as well as the iCalendar-files of the single usergroups are
loaded one every 24 hours into files on the filesystem. those files are
then used for information retrieval

3. Slight optimization of UG-display.
For that I've changed display-style from an unordered to a definition
list. I've also set the date before the meeting-information as that
visually calms the whole list a bit.