move $URI.php test higher up in error.php (phpweb bug #67871)
This test was below (way below) the $URI = strtolower($URI) line, so
when browsing to /manual/pt_BR/function.strpos we were looking for
/manual/pt_br/function.strpos.php which doesn't exist (on case-sensitive
filesystems).

Note: This is a quick fix... if it causes (significant) issues, we might
have to think of a more involved solution than shuffling the code up the
file by a few hundred lines.