Dun't recoorse-a intu iffremes. Bork Bork Bork!
Sume-a Chrume-a ixtenseeuns ooperete-a by injecteeng iffremes intu zee coorrent
ducooment. Seence-a zeese-a cume-a frum deefffferent ooreegins, ettempteeng tu
iterete-a oofer zeem in jQooery cooses a seme-a ooreegin ixcepshun tu be-a
throon. Bork Bork Bork!

Ve-a'll joost ixcloode-a iffremes, seence-a ve-a dun't use-a zeem oon zee
seete-a. Bork Bork Bork!