Adds a way to add a website to confs/cfps
This Commit adds a way to add a website-URI to a conf or a cfp-entry.
That way it is possible to link to the conferences or CfPs website
directly from the php.net website.

It should also be possible to get the link (and more informations) via a
kind of API to allow others (like joind.in or php.ug (-; ) to get that
informations