Remove the "eol" hackery that I added a couple of years ago.
For "modern" PHP branches (5.3 onwards), we'll base whether they're considered
active purely off the support dates for the branch. To ensure that we don't
immediately remove the download links for EOL branches from the front page,
we'll still consider them "active" for those purposes for 28 days after the
final release.

Basically, rather than having two sources of truth for whether a branch is EOL
or not (the flag and the support dates in branches.inc), we now have one (the
support dates). This should be an improvement.