Don't show recently EOLed branches in the supported versions table.
Changed the behaviour of get_active_branches() so that callers can
choose whether to get recently EOLed branches or not. (This is basically
for the supported versions page; the other callers -- namely the front
page download links and bug tracker backend -- likely still want the
previous behaviour.)

After all, it's a little weird listing 5.5 as a supported version in the
table when we just made a big deal of EOLing it!