Alters email-verification
The current email-verification is completely based on regex that is not
up-to-date any more in regard to newer TLDs. For example the current
version checks whether the TLD is between 2 and 6 chars long which fails
for ".digital" and some other longer TLDs

I've altered it so that the host-check is done by actually checking for
an MX-record so host that currently validate against the regex will fail
in future when no MX is available.

The Address-Part is still matched against a Regex.