Add .gitattributes file
This patch adds a new .gitattributes file to distinguish the end of
lines of the iCalendar files located in the repository and for Git to
not normalize and convert end of lines. The iCalendar format requires to
have CRLF line endings. All other files in the repository can have
native EOL. On *nix alike systems LF and on Windows CRLF which Git can
convert properly on checkouts.