Allow to access symlinked docs
Since `realpath()` resolves links, symlinked docs as suggested by the
"Setting up Documentation environment" page[1] are incompatible with
the router script.  Of course, the Webserver can be run without the
router script, but in that case shortcuts (e.g. `/json_decode`) won't
work.

Thus, we remove the `realpath()` resolution.  There is no more need
for the `strncmp()` check.  This suggest that we can get rid of the
`$afilename` and `$len` variables; hopefully, they are not used by any
of the included files.

[1] <http://doc.php.net/tutorial/local-setup.php>