Add PHPDoc to support linting via PhpStorm
Without these PHPDoc comments loading the web-php project into PhpStorm flags the variables, links and fragments as warnings or errors when using the Inspect Code feature. With these comments PhpStorm is able to recognize the purpose of the variables and to ignore the links and fragments it could not resolve so as to only flag items that might actually be real errors.