Fix to allow news2html to handle multiline 'module'
I needed a NEWS file to test bin/news2html and the only ones in the repo were /resources/NEWS_5_4_0_*.txt. Unfortunately, those had a few 'entries' that appeared to have no 'modules' and that caused the script to crash. So I added logic to recognize those and assign them to the 'Core' module and that allowed the script to work. As part of this, I unravelled the if..elseif..else logic to be a series of if..continue since the if statements were all in a loop and no code was run after the if..elseif..else statement.