improve local font support
previously local font support directs absolute paths while it is unknown to
me if paths of any kind are accepted for resolution of fonts with the CSS
`local()` function.

therefore adding the known parts that `local()` function resolves to font-
names on the local system (with preference over previous).

for me this is a great improvement as the fonts don't need to be loaded
via HTTP any longer but just work when they are locally installed using
a desktop computer.

as the Fira Sans and Fira Mono typefaces are in use on the PHP website and
my local system normally ships with these, the website loads much faster,
especially visually (fonts are used to display the text of the website).

this also reduces requests to the web-server hosting the PHP homepage
incl. the manual when the system already has the fonts.

NOTE: the circumstance that local() with absolute path-names has no effect
      is - if - not addressed in this change in so far that the original
      directives remain unchanged. this is done b/c I don't understand them
      well and might construe a flaw.