First pass at updating the homepage
In this change, I have updated the hero and navigation with a design
experiment done earlier. The changes in this commit are not final and
should not be deployed yet as the following should still be done /
checked:

- Responsiveness of the design
- Checking of all pages in the website for possible inadvertent changes
- Updating of the design on all other pages to match changes
- Update the news and conferences section to be readable in a light
background

It can be considered to add support for dark mode (this is a media query
after all); though it is out of scope for the current effort.