Modernize top half of Thanks page
In issue #682 it was requested to modernize the thanks page; as a first
step in this effort I have given the top half of the page with current
sponsors a brush up, including responsive behaviour.

Unfortunately I could not find a logo for ServerCentral; so I left that
empty.

Fixes #682