Only "linked" parameters should have pointer cursor
If a parameter has no target to scroll to, it should not suggest
otherwise by changing the cursor to a pointer.  Since the scrolling is
done via JS, we should also change the cursor via JS.[1]

While we're at it, we also fix the potential JS error where we call the
`.offset()` method on an empty `jQuery` object.

[1] <https://github.com/php/doc-en/issues/2071#issuecomment-1354497367>