Properly handle whitespace in PHP examples
PhD no longer renders NBSPs, but regular spaces instead[1].  To retain
the desired indentation, we set the CSS propery `white-space:pre`.
However, we cannot do that for the `.phpcode` class directly, because
`highlight_string()` emits additional line breaks at the top and the
bottom of the examples, so we use a more specific selector to avoid
this "padding".

[1] <https://github.com/php/phd/commit/87f794bafe22631f9886adb37db0bdf0006da3dd>